requirements.txt
flask>=2.0.0
libnmstate>=2.2.0
spacy>=3.5.0
pyyaml>=6.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
requests>=2.31.0